Shuffle source files around
Makes MessageBase::HandleClientMessage() a
protected member, as it doesn't need to be
exposed to any server code, or used externally from the message reader.

Moves BaseMessage.[cpp|h] to MessageBase,
as the type name and file name were disjoint.

Adds a basic ratelimit class for later.
We'll need a store for per-IP metadata which is collected every so often for it to be useful.