ls: claw out LS_USE_SYSTEM_EXECUTOR,...
Broken, and it's probably not gonna be in NetTS anyways, so, remove the support for it.

Removes MessageReader entirely; I think I'm gonna have the server store groups, which will effectively just be a vector<FourCC> of type codes the port accepts (could be a range too, for better concat support?)

Removes the server argument from MessageBase::HandleClientMessage; can now use Client::GetServer to get it, which IMO is nicer.

Move message reading logic to the client, there's not really a place for it to make sense elsewhere.

Document some previously undocumented things.